Moved entry point to a new Program.cs
